#78 page 컴포넌트 내 클라이언트 컴포넌트 분리
- create-view 및 일부 페이지 컴포넌트들 중 클라이언트 컴포넌트인 요소들은 별도 container 하위 클라이언트 컴포넌트로 분리
- page 컴포넌트들은 server 컴포넌트로 유지
